requirements.txt
Django==6.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
django-cors-headers==4.7.0
django-crontab==0.7.1

mysqlclient==2.2.7

python-dotenv==1.1.1
requests==2.32.3

twilio==9.7.1

pillow==11.2.1
reportlab==4.4.1
xhtml2pdf==0.2.17
